Reverted last commit - 1.1.5
The last release was just buggy all around due to me currently rewriting
Vixio and not messing with that code, I found it had a few new issues so
I reverted it back to 1.1.3 and just fixed the login issue caused by
Discord and the returning of incorrect values. You can follow the
rewrite in the rewrite branch of the Vixio repo! If you have issues on
startup with this build try deleting your Vixio folder and restarting
the server.