Add regex check to role named
The roles named expression now returns true in isSingle if you dont include the s, this means you do not have to set a list variable to the result if you know there is only one role and use `role` instead of `roles`. It will return the first role by the name if you dont include the `s` and there are multiple roles with the same name.