Rework changer system to allow bots to be assumed
Now in events with bots, you no longer must specify the bot! So things like `add role with id "2312132" to roles of event-member` no longer need you to specify the bot and will use the one from the event! Large thanks to @pikachu920 for the help with this addition, one large bug down!